move time-service out of platform
