+++ src/CMakeLists.txt
@@ -109,7 +109,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/platform/hid-device-info.h"
         "${CMAKE_CURRENT_LIST_DIR}/platform/playback-device-info.h"
         "${CMAKE_CURRENT_LIST_DIR}/platform/stream-profile.h"
-        "${CMAKE_CURRENT_LIST_DIR}/platform/time-service.h"
+        "${CMAKE_CURRENT_LIST_DIR}/core/time-service.h"
         "${CMAKE_CURRENT_LIST_DIR}/platform/uvc-device.h"
         "${CMAKE_CURRENT_LIST_DIR}/platform/uvc-device-info.h"
         "${CMAKE_CURRENT_LIST_DIR}/context.h"
