+++ src/environment.cpp
@@ -245,12 +245,12 @@ namespace librealsense
         return _extrinsics;
     }
 
-    void environment::set_time_service(std::shared_ptr<platform::time_service> ts)
+    void environment::set_time_service(std::shared_ptr<time_service> ts)
     {
         _ts = ts;
     }
 
-    std::shared_ptr<platform::time_service> environment::get_time_service()
+    std::shared_ptr<time_service> environment::get_time_service()
     {
         return _ts;
     }
