+++ src/linux/backend-v4l2.cpp
@@ -3,7 +3,7 @@
 
 #include "backend-v4l2.h"
 #include <src/platform/command-transfer.h>
-#include <src/platform/time-service.h>
+#include <src/core/time-service.h>
 #include "backend-hid.h"
 #include "backend.h"
 #include "types.h"
