+++ wrappers/python/pybackend.cpp
@@ -17,7 +17,7 @@ Copyright(c) 2017 Intel Corporation. All Rights Reserved. */
 #include "core/options.h"   // Workaround for the missing DLL_EXPORT template
 #include "core/info.h"   // Workaround for the missing DLL_EXPORT template
 #include "../src/backend.h"
-#include <src/platform/time-service.h>
+#include <src/core/time-service.h>
 #include <src/platform/command-transfer.h>
 #include <src/platform/hid-device.h>
 #include "pybackend_extras.h"
@@ -61,10 +61,10 @@ PYBIND11_MODULE(NAME, m) {
                  .def_readwrite("def", &platform::control_range::def)
                  .def_readwrite("step", &platform::control_range::step);
 
-    py::class_<platform::time_service> time_service(m, "time_service");
-    time_service.def("get_time", &platform::time_service::get_time);
+    py::class_<librealsense::time_service> time_service(m, "time_service");
+    time_service.def("get_time", &librealsense::time_service::get_time);
 
-    py::class_<platform::os_time_service, platform::time_service> os_time_service(m, "os_time_service");
+    py::class_<librealsense::os_time_service, librealsense::time_service> os_time_service(m, "os_time_service");
 
 #define BIND_RAW_RO_ARRAY(class, name, type, size) #name, [](const class &c) -> const std::array<type, size>& { return reinterpret_cast<const std::array<type, size>&>(c.name); }
 #define BIND_RAW_RW_ARRAY(class, name, type, size) BIND_RAW_RO_ARRAY(class, name, type, size), [](class &c, const std::array<type, size> &arr) { for (int i=0; i<size; ++i) c.name[i] = arr[i]; }
